+++ test/src/tests/tcp.py
@@ -14,7 +14,7 @@
 # You should have received a copy of the GNU General Public License
 # along with Masscanned. If not, see <http://www.gnu.org/licenses/>.
 
-from scapy.layers.inet import IP, ICMP, TCP
+from scapy.layers.inet import IP, TCP
 from scapy.layers.inet6 import IPv6
 from scapy.layers.l2 import Ether
 from scapy.packet import Raw
@@ -23,34 +23,6 @@ from scapy.volatile import RandInt
 
 from ..conf import IPV4_ADDR, IPV6_ADDR, MAC_ADDR
 from ..core import test, check_ip_checksum, check_ipv6_checksum
-
-
-@test
-def test_ipv4_req():
-    ##### IP #####
-    ip_req = Ether(dst=MAC_ADDR) / IP(dst=IPV4_ADDR, id=0x1337) / ICMP(type=8, code=0)
-    ip_repl = srp1(ip_req, timeout=1)
-    assert ip_repl is not None, "expecting answer, got nothing"
-    check_ip_checksum(ip_repl)
-    assert IP in ip_repl, "no IP layer in response"
-    ip_repl = ip_repl[IP]
-    assert ip_repl.id == 0, "IP identification unexpected"
-
-
-@test
-def test_eth_req_other_mac():
-    #### ETH ####
-    ip_req = Ether(dst="00:00:00:11:11:11") / IP(dst=IPV4_ADDR) / ICMP(type=8, code=0)
-    ip_repl = srp1(ip_req, timeout=1)
-    assert ip_repl is None, "responding to other MAC addresses"
-
-
-@test
-def test_ipv4_req_other_ip():
-    ##### IP #####
-    ip_req = Ether(dst=MAC_ADDR) / IP(dst="1.2.3.4") / ICMP(type=8, code=0)
-    ip_repl = srp1(ip_req, timeout=1)
-    assert ip_repl is None, "responding to other IP addresses"
 
 
 @test
